third_party/WebKit/LayoutTests/http/tests/inspector/security/mixed-content-reload-expected.txt

Tests that the mixed content explanation prompts the user to refresh when there are no recorded requests, and links to the network panel when there are recorded requests.

<DIV class=security-explanation security-explanation-neutral >
    <DIV class=security-property security-property-neutral >
    </DIV>
    <DIV class=security-explanation-text >
        <DIV class=security-explanation-title >
Mixed Content
        </DIV>
        <DIV >
The site includes HTTP resources.
        </DIV>
        <DIV class=security-mixed-content >
Reload the page to record requests for HTTP resources.
        </DIV>
    </DIV>
</DIV>
<DIV class=security-explanation security-explanation-neutral >
    <DIV class=security-property security-property-neutral >
    </DIV>
    <DIV class=security-explanation-text >
        <DIV class=security-explanation-title >
Mixed Content
        </DIV>
        <DIV >
The site includes HTTP resources.
        </DIV>
        <DIV class=security-mixed-content link >
View 1 request in Network Panel
        </DIV>
    </DIV>
</DIV>
